ai_backend/requirements.txt
fastapi
uvicorn[gunicorn]
openai  # or your own model
pydantic
motor
python-dotenv
requests
aiofiles
httpx